// src/abstract/api/user/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Hospital Management Service
 * 
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface UserApi {
	getProfile: (id: string | undefined) => Promise<GetProfileResponse>
	updatepatient: (id: string | undefined) => Promise<UpdatepatientResponse>
	signIn: (request: Api.SignInRequest | undefined) => Promise<SignInResponse>
}

export type GetProfileResponse = GetProfile200Response | GetProfile404Response

export interface GetProfile200Response {
	status: 200
	body: Api.GetProfile200Response
	headers?: never
}

export interface GetProfile404Response {
	status: 404
	body: Api.GetProfile404Response
	headers?: never
}

export type UpdatepatientResponse = Updatepatient200Response | Updatepatient404Response

export interface Updatepatient200Response {
	status: 200
	body: Api.Updatepatient200Response
	headers?: never
}

export interface Updatepatient404Response {
	status: 404
	body: Api.Updatepatient404Response
	headers?: never
}

export type SignInResponse = SignIn200Response | SignIn404Response | SignIn500Response

export interface SignIn200Response {
	status: 200
	body: Api.SignIn200Response
	headers?: never
}

export interface SignIn404Response {
	status: 404
	body: Api.SignIn404Response
	headers?: never
}

export interface SignIn500Response {
	status: 500
	body: Api.SignIn500Response
	headers?: never
}